re-organised my filter files and reworked sortBy F
