+++ index.html
@@ -10,6 +10,7 @@
     <link rel="stylesheet" href="/src/assets/fontawsome/css/all.css">
     <script type="module" src="/src/js/index.mjs"></script>
     <script type="module" src="/src/js/posts/posts.mjs"></script>
+    <script type="module" src="/src/js/filter/oldest-newest.mjs"></script>
 </head>
 <body>
     <header class="text-bg-primary">
