+++ src/js/index.mjs
@@ -1,7 +1,7 @@
 import { logout } from "./menu-items/logout.mjs";
 import { logInOutIndicator } from "./menu-items/login-out-indicator.mjs";
 import { searchHandler } from "./search/search.mjs";
-import { filterHandler } from "./filter/filter.mjs";
+import { tagHandler } from "./filter/tags.mjs";
 
 logout();
 logInOutIndicator();
